LayoutTests/imported/w3c/web-platform-tests/css/css-scrollbars/scrollbar-width-004-expected.txt

Test scrollbar-width: both scrollbars, vertical-lr
auto
thin
none
Test scrollbar-width: both scrollbars, vertical-rl
auto
thin
none

PASS vertical-lr, scrollbar-width auto
PASS vertical-lr, scrollbar-width thin
FAIL vertical-lr, scrollbar-width none assert_equals: ltr auto clientWidth expected 200 but got 185
PASS vertical-rl, scrollbar-width auto
PASS vertical-rl, scrollbar-width thin
FAIL vertical-rl, scrollbar-width none assert_equals: rtl auto clientWidth expected 200 but got 185
